+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  - Add options for relative effect locations (Thanks, SexyToad!)
  - Add "track_range" CustomProjectile parameter, for projectiles that track the player's cursor at a fixed range (Thanks again, SexyToad!)
  - Add Railgun spell, ability for CustomProjectile to hit multiple targets  
+ - Add mathematically-controlled custom projectile velocity (Thanks yet again, SexyToad!)
 
 ## 5.4